Move "bugs.go" out of util and into its own package
