+++ oneCRL/oneCRL.go
@@ -14,8 +14,8 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/mozilla/OneCRL-Tools/bugs"
 	"github.com/mozilla/OneCRL-Tools/config"
-	"github.com/mozilla/OneCRL-Tools/util"
 	"io/ioutil"
 	"log"
 	"net/http"
